// client/src/components/mood/dashboard/trends-card/trends-card.module.scss.d.ts
export type Styles = {
  'trendsCard': string;
  'trendsCard_Chart': string;
  'trendsCard_Chart_Xaxis': string;
  'trendsCard_Chart_Yaxis': string;
  'trendsCard_Chart_Yaxis_Item': string;
  'trendsCard_Heading': string;
  'trendsCard_MonthPicker': string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;